requirements.txt
fastapi==0.110.2
pydantic==2.7.1
uvicorn==0.29.0
tabula-py==2.9.0
requests==2.31.0
